Changed Integer to Double.

+++ src/AI/Algorithms/Minimax.java
@@ -33,7 +33,7 @@ final public class Minimax implements IMinimax{
 		for (int i = 0; i < actions.size(); i++){
 			TicTacToe newState = state.deepClone();
 			newState.move(actions.get(i));
-            double min = minValue(newState, depthP - 1);
+            		double min = minValue(newState, depthP - 1);
 			if (min > v){
 				v = min;
 			}
@@ -64,7 +64,7 @@ final public class Minimax implements IMinimax{
 	@Override
 	public AbsMove computeMove(AbsTicTacToe state){
 		ArrayList<AbsMove> actions = AIUtils.computeActions(state.getField());
-		double max = Integer.MIN_VALUE;
+		double max = Double.MIN_VALUE;
 		AbsMove bestMove = null;
 		
 		for (int i = 0; i < actions.size(); i++){
